requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.0
pydantic>=2.5.0
pytest>=7.4.0
httpx>=0.25.0
python-multipart>=0.0.6
pytest-asyncio>=0.21.0
asgiref>=3.7.2
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6